backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask_SQLAlchemy==3.0.2
gunicorn==21.2.0
joblib==1.3.2
psycopg2-binary==2.9.9
python-dotenv=1.0.0
PyYAML==6.0.1
qrcode==7.4.2
scikit-learn==1.3.2